requirements.txt
torch==2.7.0
numpy==2.2.5
pandas==2.2.3
scikit-learn==1.6.1
h5py==3.13.0
tqdm==4.67.1
pillow==11.2.1
matplotlib==3.10.1
torchvision==0.22.0
hydra-core==1.3.2
omegaconf==2.3.0
wandb==0.19.10